Option to not auto discover maven root
In multi maven projects it can be helpful to assume that the directoy
the editor has been set to can be assumed to be the maven root since
otherwise only a subproject would be built causing build failures where
there should not be any.

This is enabled via the `g:maven_detect_root` which defaults to `1` for
true but can be set to `0` making the current directory the maven root.